requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
shap>=0.43
flask>=3.0
joblib>=1.3
matplotlib>=3.7
seaborn>=0.13
imbalanced-learn>=0.11
pytest>=7.4
python-dotenv>=1.0
